+++ docs/.vitepress/theme/index.ts
@@ -19,6 +19,9 @@ import CodePreview from '../../../src/app_components/CodePreview.vue';
 import '../../../style/message';
 import '../../../style/shadow';
 import '../../../style/table';
+import '../../../style/switch';
+import '../../../style/form';
+import '../../../style/form_item';
 
 export default {
   extends: DefaultTheme,